fix(odk): remove first level (None) in submissions payload
