+++ aether-odk-module/aether/odk/api/tests/test_xform_utils.py
@@ -2,7 +2,7 @@ from django.core.exceptions import ValidationError
 
 from . import CustomTestCase
 
-from ..xform_utils import validate_xmldict
+from ..xform_utils import validate_xmldict, extract_data_from_xml, parse_submission
 
 
 class XFormUtilsTests(CustomTestCase):
@@ -84,3 +84,16 @@ class XFormUtilsTests(CustomTestCase):
                 </h:html>
             '''
         )
+
+    def test__parse_submission(self):
+        with open(self.samples['submission']['file-ok'], 'rb') as xml:
+            data, form_id, version = extract_data_from_xml(xml)
+
+        self.assertEqual(form_id, 'my-test-form')
+        self.assertEqual(version, '0')
+        self.assertEqual(len(list(data.keys())), 1)
+        self.assertEqual(list(data.keys())[0], 'Something_that_is_not_None')
+
+        data = parse_submission(data, self.samples['xform']['raw-xml'])
+
+        self.assertNotEqual(list(data.keys())[0], 'Something_that_is_not_None')
